+++ cli/aux/opts.c
@@ -23,6 +23,13 @@ void tec_cli_rm_option_init(struct tec_cli_rm_options *opts)
     opts->mode = RMI_ALWAYS;
 }
 
+void tec_cli_ls_option_init(struct tec_cli_ls_options *opts)
+{
+    opts->help = false;
+    opts->quiet = false;
+    opts->togg = false;
+}
+
 void tec_cli_set_option_init(struct tec_cli_set_options *opts)
 {
     opts->help = false;